Move fixture macsec_setup to macsec/__init__.py
